backend/requirements.txt
pytest
pydantic
requests
python-dotenv
openai-agents
fastapi
uvicorn
